Restructure data generation module
